requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
colorama==0.4.4
DateTime==4.3
dnspython==2.1.0
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zope.interface==5.4.0
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
colorama==0.4.4
DateTime==4.3
dnspython==2.1.0
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyMySQL==1.0.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zope.interface==5.4.0
